Ask for more bits
Explain how to avoid an `EE_KEY_TOO_SMALL` error by requesting a larger proxy.
(This error may pop up on newer OSs depending on the OpenSSL config or system-wide crypto policy.)
Technically it's not required on moria but it's harmless and I figured I'd go for consistency.
I couldn't find a way to ask cigetcert/kx509 for more bits but they already make 2048-bit proxies so we're good.